requirements.txt
Flask
Pillow
SQLAlchemy
redis
rq
locust